python/arch_diagram_parser_local_controller/requirements.txt
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib